refactor(tools/gltf_auto_export): complete restructuring of the tool  * split up code in logical modules instead of one giant file  * added module structure, imports etc  * updated README
