+++ .gitignore
@@ -1,2 +1,3 @@
 *.blend1
 target/
+__pycache__/
