+++ tools/gltf_auto_export/__init__.py
@@ -1,8 +1,7 @@
-#TODO: this is not actually in use yet, just use the gltf_auto_export.py file
 bl_info = {
-    "name": "Test glTF/glb auto-export",
+    "name": "gltf_auto_export",
     "author": "kaosigh",
-    "version": (0, 1),
+    "version": (0, 5, 4),
     "blender": (3, 4, 0),
     "location": "File > Import-Export",
     "description": "glTF/glb auto-export",
@@ -10,37 +9,213 @@ bl_info = {
     "wiki_url": "",
     "tracker_url": "",
     "category": "Import-Export"
-    }
-
+}
 import bpy
+import os
 
-from .gltf_auto_export import TEST_AUTO_OT_gltf
-from .gltf_auto_export import deps_update_handler
-from .gltf_auto_export import save_handler
-from .gltf_auto_export import get_changedScene
-from .gltf_auto_export import set_ChangedScene
+from bpy.app.handlers import persistent
+from bpy.props import (BoolProperty,
+                       IntProperty,
+                       StringProperty,
+                       EnumProperty,
+                       CollectionProperty
+                       )
 
 
-# Only needed if you want to add into a dynamic menu
+from . import helpers
+from .internals import (SceneLink,
+                        SceneLinks,
+                        CollectionToExport,
+                        CollectionsToExport,
+                        CUSTOM_PG_sceneName
+                        )
+from .auto_export import auto_export
+from .preferences import (AutoExportGltfPreferenceNames,
+                          AutoExportGltfAddonPreferences
+                          )
+from .ui.main import (GLTF_PT_auto_export_main,
+                      GLTF_PT_auto_export_root,
+                      GLTF_PT_auto_export_blueprints,
+                      GLTF_PT_auto_export_collections_list,
+                      GLTF_PT_auto_export_gltf,
+                      GLTF_auto_export_UL_SCENES,
+                      AutoExportGLTF
+                      )
+from .ui.various import (CUSTOM_OT_actions)
+from .helpers_scenes import (is_scene_ok)
+
+bpy.context.window_manager['changed_objects_per_scene'] = {}
+bpy.context.window_manager['previous_params'] = {}
+bpy.context.window_manager['__gltf_auto_export_initialized'] = False
+bpy.context.window_manager['__gltf_auto_export_gltf_params_changed'] = False
+
+class TESTADDON_PT_TestPanel(bpy.types.Panel):
+    bl_idname = "TESTADDON_PT_TestPanel"
+    bl_label = "Test Addon"
+    bl_space_type = 'VIEW_3D'
+    bl_region_type = 'UI'
+    bl_category = "Test Addon"
+    bl_context = "objectmode"
+
+    def draw(self, context):
+
+        layout = self.layout
+
+        row = layout.row()
+        row.label(text="How cool is this!AUTO_export"  )
+
+######################################################
+
+
+
+#see here for original gltf exporter infos https://github.com/KhronosGroup/glTF-Blender-IO/blob/main/addons/io_scene_gltf2/__init__.py
+@persistent
+def deps_update_handler(scene, depsgraph):
+    if scene.name != "temp_scene": # actually do we care about anything else than the main scene(s) ?
+        print("depsgraph_update_post", scene.name)
+        print("-------------")
+        changed = scene.name or ""
+
+        # depsgraph = bpy.context.evaluated_depsgraph_get()
+        if not 'changed_objects_per_scene' in bpy.context.window_manager:
+            bpy.context.window_manager['changed_objects_per_scene'] = {}
+
+        if not changed in bpy.context.window_manager['changed_objects_per_scene']:
+            bpy.context.window_manager['changed_objects_per_scene'][changed] = {}
+        
+        for obj in depsgraph.updates:
+            if isinstance(obj.id, bpy.types.Object):
+                # get the actual object
+                object = bpy.data.objects[obj.id.name]
+                bpy.context.window_manager['changed_objects_per_scene'][scene.name][obj.id.name] = object
+        
+        bpy.context.window_manager.changedScene = changed
+
+@persistent
+def save_handler(dummy): 
+    print("-------------")
+    print("saved", bpy.data.filepath)
+    if not 'changed_objects_per_scene' in bpy.context.window_manager:
+        bpy.context.window_manager['changed_objects_per_scene'] = {}
+    changes_per_scene =  bpy.context.window_manager['changed_objects_per_scene']
+
+    #determine changed parameters
+    addon_prefs = bpy.context.preferences.addons["gltf_auto_export"].preferences
+
+    prefs = {}
+    for (k,v) in addon_prefs.items():
+        if k not in AutoExportGltfPreferenceNames:
+            prefs[k] = v
+
+    previous_params = bpy.context.window_manager['previous_params'] if 'previous_params' in bpy.context.window_manager else {}
+    set1 = set(previous_params.items())
+    set2 = set(prefs.items())
+    difference = dict(set1 ^ set2)
+    
+    changed_param_names = list(set(difference.keys())- set(AutoExportGltfPreferenceNames))
+    changed_parameters = len(changed_param_names) > 0
+    # do the export
+    auto_export(changes_per_scene, changed_parameters)
+
+
+    # save the parameters
+    # todo add back
+    for (k, v) in prefs.items():
+        bpy.context.window_manager['previous_params'][k] = v
+
+    # reset a few things after exporting
+    # reset wether the gltf export paramters were changed since the last save 
+    bpy.context.window_manager['__gltf_auto_export_gltf_params_changed'] = False
+    # reset whether there have been changed objects since the last save 
+    bpy.context.window_manager['changed_objects_per_scene'] = {}
+
+def get_changedScene(self):
+    return self["changedScene"]
+
+def set_changedScene(self, value):
+    self["changedScene"] = value
+
+classes = [
+    SceneLink,
+    SceneLinks,
+    CUSTOM_PG_sceneName,
+    GLTF_auto_export_UL_SCENES,
+    CUSTOM_OT_actions,
+
+    AutoExportGLTF, 
+    AutoExportGltfAddonPreferences,
+
+    CollectionToExport,
+    CollectionsToExport,
+
+    GLTF_PT_auto_export_main,
+    GLTF_PT_auto_export_root,
+    GLTF_PT_auto_export_blueprints,
+    GLTF_PT_auto_export_collections_list,
+    GLTF_PT_auto_export_gltf
+]
+
 def menu_func_import(self, context):
-    self.layout.operator(TEST_AUTO_OT_gltf.bl_idname, text="glTF auto Export (.glb/gltf)")
-
+    self.layout.operator(AutoExportGLTF.bl_idname, text="glTF auto Export (.glb/gltf)")
 
 def register():
-    bpy.utils.register_class(TEST_AUTO_OT_gltf)
-    bpy.types.TOPBAR_MT_file_export.append(menu_func_import)
+    bpy.utils.register_class(TESTADDON_PT_TestPanel)
+    for cls in classes:
+        bpy.utils.register_class(cls)
 
+    bpy.types.Scene.main_scene = bpy.props.PointerProperty(type=bpy.types.Scene, name="main scene", description="main_scene_chooser", poll=is_scene_ok)
+    bpy.types.Scene.library_scene = bpy.props.PointerProperty(type=bpy.types.Scene, name="library scene", description="library_scene_picker", poll=is_scene_ok)
+
+    # setup handlers for updates & saving
     bpy.app.handlers.depsgraph_update_post.append(deps_update_handler)
     bpy.app.handlers.save_post.append(save_handler)
-    #bpy.types.TOPBAR_MT_file_export.append(menu_func_import)
-    bpy.types.Scene.changedScene = bpy.props.StringProperty(get=get_changedScene, set=set_ChangedScene)
 
+    bpy.types.WindowManager.changedScene = bpy.props.StringProperty(get=get_changedScene, set=set_changedScene)
+    bpy.types.WindowManager.exportedCollections = bpy.props.CollectionProperty(type=CollectionsToExport)
+
+    # add our addon to the toolbar
+    bpy.types.TOPBAR_MT_file_export.append(menu_func_import)
+
+    ## just experiments
+    bpy.types.Scene.main_scenes_list_index = IntProperty(name = "Index for main scenes list", default = 0)
+    bpy.types.Scene.library_scenes_list_index = IntProperty(name = "Index for library scenes list", default = 0)
+
+    
+    mock_main_scenes = []#["World", "level2"]
+    main_scenes = bpy.context.preferences.addons["gltf_auto_export"].preferences.main_scenes
+    for item_name in mock_main_scenes:
+        item = main_scenes.add()
+        item.name = item_name
+    
+    mock_library_scenes = [] #["Library", "Library2"]
+    library_scenes = bpy.context.preferences.addons["gltf_auto_export"].preferences.library_scenes
+    for item_name in mock_library_scenes:
+        item = library_scenes.add()
+        item.name = item_name
+
+    bpy.context.preferences.addons["gltf_auto_export"].preferences.main_scenes_index = 0
+    bpy.context.preferences.addons["gltf_auto_export"].preferences.library_scenes_index = 0
 
 def unregister():
-    bpy.utils.unregister_class(TEST_AUTO_OT_gltf)
+    bpy.utils.unregister_class(TESTADDON_PT_TestPanel)
+    for cls in classes:
+        bpy.utils.unregister_class(cls)
+
     bpy.types.TOPBAR_MT_file_export.remove(menu_func_import)
 
+    # remove handlers & co
     bpy.app.handlers.depsgraph_update_post.remove(deps_update_handler)
     bpy.app.handlers.save_post.remove(save_handler)
-    #bpy.types.TOPBAR_MT_file_export.remove(menu_func_import)
-    del bpy.types.Scene.changedScene
+    
+    del bpy.types.WindowManager.changedScene
+    del bpy.types.WindowManager.exportedCollections
+
+    del bpy.types.Scene.main_scene
+    del bpy.types.Scene.library_scene
+
+    del bpy.types.Scene.main_scenes_list_index
+    del bpy.types.Scene.library_scenes_list_index
+
+
+if "gltf_auto_export" == "__main__":
+    register()
