+++ tools/gltf_auto_export/config.py
@@ -1,0 +1,1 @@
+scene_key = "auto_gltfExportSettings"
